chore: create only one handle for all of proposal types
