+++ components/Members/AddMemberForm.tsx
@@ -1,4 +1,3 @@
-import useMembersListStore from 'stores/useMembersStore'
 import { PublicKey } from '@solana/web3.js'
 import useRealm from 'hooks/useRealm'
 import Input from 'components/inputs/Input'
@@ -11,51 +10,49 @@ import { getMintSchema } from 'utils/validations'
 import { useEffect, useState } from 'react'
 import { MintForm, UiInstruction } from 'utils/uiTypes/proposalCreationTypes'
 import useGovernanceAssets from 'hooks/useGovernanceAssets'
-import { getInstructionDataFromBase64 } from 'models/serialisation'
-import { RpcContext } from 'models/core/api'
 import { Governance } from 'models/accounts'
 import { ParsedAccount } from 'models/core/accounts'
-import { useRouter } from 'next/router'
-import { createProposal } from 'actions/createProposal'
-import { notify } from 'utils/notifications'
-import useQueryContext from 'hooks/useQueryContext'
 import { getMintInstruction } from 'utils/instructionTools'
 import AddMemberIcon from '@components/AddMemberIcon'
-import Tooltip from '@components/Tooltip'
 import {
   ArrowCircleDownIcon,
   ArrowCircleUpIcon,
 } from '@heroicons/react/outline'
+import TokenBalanceCard from '@components/TokenBalanceCard'
+import { handlePropose } from 'actions/handleCreateProposal'
+import { useRouter } from 'next/router'
+import useQueryContext from '@hooks/useQueryContext'
+import { notify } from '@utils/notifications'
 
 interface AddMemberForm extends MintForm {
   description: string
   title: string
 }
 
-const AddMemberForm = ({ close }) => {
+const AddMemberForm = ({
+  close,
+  callback,
+}: {
+  close?: any
+  callback?: any
+}) => {
   const [voteByCouncil, setVoteByCouncil] = useState(false)
   const [showOptions, setShowOptions] = useState(false)
   const [isLoading, setIsLoading] = useState(false)
   const [formErrors, setFormErrors] = useState({})
-  const [lastStep, setLastStep] = useState(false)
 
-  const router = useRouter()
   const connection = useWalletStore((s) => s.connection)
   const wallet = useWalletStore((s) => s.current)
 
-  const { fmtUrlWithCluster } = useQueryContext()
   const { fetchRealmGovernance } = useWalletStore((s) => s.actions)
-  const { symbol } = router.query
   const { getMintWithGovernances } = useGovernanceAssets()
 
-  const {
-    realmInfo,
-    canChooseWhoVote,
-    councilMint,
-    realm,
-    ownVoterWeight,
-    mint,
-  } = useRealm()
+  const realmData = useRealm()
+
+  const router = useRouter()
+  const { fmtUrlWithCluster } = useQueryContext()
+
+  const { realmInfo, canChooseWhoVote, councilMint, realm } = realmData
 
   const programId: PublicKey | undefined = realmInfo?.programId
 
@@ -68,8 +65,6 @@ const AddMemberForm = ({ close }) => {
     title: '',
   })
 
-  const schema = getMintSchema({ form, connection })
-
   const mintMinAmount = form.mintAccount
     ? getMintMinAmountAsDecimal(councilMint!)
     : 1
@@ -77,17 +72,6 @@ const AddMemberForm = ({ close }) => {
   const currentPrecision = precision(mintMinAmount)
 
   const proposalTitle = `Add council member ${form.destinationAccount}`
-
-  const defaultProposalMint = !mint?.supply.isZero()
-    ? realm?.info.communityMint
-    : !councilMint?.supply.isZero()
-    ? realm?.info.config.councilMint
-    : undefined
-
-  const proposalMint =
-    canChooseWhoVote && voteByCouncil
-      ? realm?.info.config.councilMint
-      : defaultProposalMint
 
   const setAmount = (event) => {
     const value = event.target.value
@@ -117,6 +101,8 @@ const AddMemberForm = ({ close }) => {
     })
   }
 
+  const schema = getMintSchema({ form, connection })
+
   const getInstruction = async (): Promise<UiInstruction> => {
     return getMintInstruction({
       schema,
@@ -129,92 +115,10 @@ const AddMemberForm = ({ close }) => {
     })
   }
 
-  //TODO common handle propose
-  const handlePropose = async () => {
-    setIsLoading(true)
-
-    const instruction: UiInstruction = await getInstruction()
-
-    if (instruction.isValid) {
-      const governance = form.mintAccount?.governance
-
-      let proposalAddress: PublicKey | null = null
-
-      if (!realm) {
-        setIsLoading(false)
-
-        throw new Error('No realm selected')
-      }
-
-      const rpcContext = new RpcContext(
-        new PublicKey(realm.account.owner.toString()),
-        realmInfo?.programVersion,
-        wallet,
-        connection.current,
-        connection.endpoint
-      )
-
-      const instructionData = {
-        data: instruction.serializedInstruction
-          ? getInstructionDataFromBase64(instruction.serializedInstruction)
-          : null,
-        holdUpTime: governance?.info?.config.minInstructionHoldUpTime,
-        prerequisiteInstructions: instruction.prerequisiteInstructions || [],
-      }
-
-      try {
-        const selectedGovernance = (await fetchRealmGovernance(
-          governance?.pubkey
-        )) as ParsedAccount<Governance>
-
-        const ownTokenRecord = ownVoterWeight.getTokenRecordToCreateProposal(
-          governance!.info.config
-        )
-
-        const defaultProposalMint = !mint?.supply.isZero()
-          ? realm.info.communityMint
-          : !councilMint?.supply.isZero()
-          ? realm.info.config.councilMint
-          : undefined
-
-        const proposalMint =
-          canChooseWhoVote && voteByCouncil
-            ? realm.info.config.councilMint
-            : defaultProposalMint
-
-        if (!proposalMint) {
-          throw new Error(
-            'There is no suitable governing token for the proposal'
-          )
-        }
-
-        proposalAddress = await createProposal(
-          rpcContext,
-          realm.pubkey,
-          selectedGovernance.pubkey,
-          ownTokenRecord.pubkey,
-          form.title ? form.title : proposalTitle,
-          form.description ? form.description : '',
-          proposalMint,
-          selectedGovernance?.info?.proposalCount,
-          [instructionData],
-          false
-        )
-
-        const url = fmtUrlWithCluster(
-          `/dao/${symbol}/proposal/${proposalAddress}`
-        )
-
-        router.push(url)
-      } catch (error) {
-        notify({
-          type: 'error',
-          message: `${error}`,
-        })
-      }
-    }
-
-    setIsLoading(false)
+  const getSelectedGovernance = async () => {
+    return (await fetchRealmGovernance(
+      form.mintAccount?.governance?.pubkey
+    )) as ParsedAccount<Governance>
   }
 
   useEffect(() => {
@@ -234,134 +138,265 @@ const AddMemberForm = ({ close }) => {
     initForm()
   }, [])
 
+  const modalCallback = ({ url, error }) => {
+    let newAddressURL = null
+
+    if (url) {
+      newAddressURL = fmtUrlWithCluster(url)
+    }
+
+    if (close) {
+      if (error) {
+        close()
+
+        notify({
+          type: 'error',
+          message: 'Could not create proposal',
+          description: `${error}`,
+        })
+
+        return
+      }
+
+      router.push(String(newAddressURL))
+
+      return
+    }
+  }
+
+  const confirmPropose = async () => {
+    return await handlePropose({
+      getInstruction,
+      form,
+      connection,
+      callback: close ? modalCallback : callback,
+      governance: form.mintAccount?.governance,
+      realmData,
+      wallet,
+      getSelectedGovernance,
+      setIsLoading,
+    })
+  }
+
   return (
     <>
-      {close && (
-        <div className="flex justify-start items-center gap-x-3">
-          <AddMemberIcon className="w-8 mb-2" />
-
-          <h2 className="text-xl">
-            Add new member to {realmInfo?.displayName}
-          </h2>
-        </div>
-      )}
-
-      <Input
-        useDefaultStyle={false}
-        className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
-        wrapperClassName={`${close ? 'mt-0 mb-6' : 'my-6'}`}
-        label="Member's wallet"
-        placeholder="Member's wallet"
-        value={form.destinationAccount}
-        type="text"
-        onChange={(event) =>
-          handleSetForm({
-            value: event.target.value,
-            propertyName: 'destinationAccount',
-          })
-        }
-        noMaxWidth
-        error={formErrors['destinationAccount']}
-      />
-
-      <div
-        className={'flex items-center hover:cursor-pointer w-24 my-3'}
-        onClick={() => setShowOptions(!showOptions)}
-      >
-        {showOptions ? (
-          <ArrowCircleUpIcon className="h-4 w-4 mr-1 text-primary-light" />
-        ) : (
-          <ArrowCircleDownIcon className="h-4 w-4 mr-1 text-primary-light" />
-        )}
-        <small className="text-fgd-3">Options</small>
-      </div>
-
-      {showOptions && (
+      {close ? (
         <>
+          <div className="flex justify-start items-center gap-x-3">
+            <AddMemberIcon className="w-8 mb-2" />
+
+            <h2 className="text-xl">
+              Add new member to {realmInfo?.displayName}
+            </h2>
+          </div>
+
           <Input
-            noMaxWidth
             useDefaultStyle={false}
             className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
-            wrapperClassName="mb-6"
-            label="Title of your proposal"
-            placeholder="Title of your proposal"
-            value={form.title}
+            wrapperClassName={`${close ? 'mt-0 mb-6' : 'my-6'}`}
+            label="Member's wallet"
+            placeholder="Member's wallet"
+            value={form.destinationAccount}
             type="text"
             onChange={(event) =>
               handleSetForm({
                 value: event.target.value,
-                propertyName: 'title',
+                propertyName: 'destinationAccount',
               })
             }
-          />
-
-          <Input
             noMaxWidth
-            useDefaultStyle={false}
-            className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
-            wrapperClassName="mb-6"
-            min={mintMinAmount}
-            label="Voter weight"
-            value={form.amount}
-            type="number"
-            onChange={setAmount}
-            step={mintMinAmount}
-            error={formErrors['amount']}
-            onBlur={validateAmountOnBlur}
+            error={formErrors['destinationAccount']}
           />
 
-          {canChooseWhoVote && (
+          <div
+            className={'flex items-center hover:cursor-pointer w-24 my-3'}
+            onClick={() => setShowOptions(!showOptions)}
+          >
+            {showOptions ? (
+              <ArrowCircleUpIcon className="h-4 w-4 mr-1 text-primary-light" />
+            ) : (
+              <ArrowCircleDownIcon className="h-4 w-4 mr-1 text-primary-light" />
+            )}
+            <small className="text-fgd-3">Options</small>
+          </div>
+
+          {showOptions && (
+            <>
+              <Input
+                noMaxWidth
+                useDefaultStyle={false}
+                className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
+                wrapperClassName="mb-6"
+                label="Title of your proposal"
+                placeholder="Title of your proposal"
+                value={form.title}
+                type="text"
+                onChange={(event) =>
+                  handleSetForm({
+                    value: event.target.value,
+                    propertyName: 'title',
+                  })
+                }
+              />
+
+              <Input
+                noMaxWidth
+                useDefaultStyle={false}
+                className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
+                wrapperClassName="mb-6"
+                min={mintMinAmount}
+                label="Voter weight"
+                value={form.amount}
+                type="number"
+                onChange={setAmount}
+                step={mintMinAmount}
+                error={formErrors['amount']}
+                onBlur={validateAmountOnBlur}
+              />
+
+              {canChooseWhoVote && (
+                <VoteBySwitch
+                  disabled={!canChooseWhoVote}
+                  checked={voteByCouncil}
+                  onChange={() => {
+                    setVoteByCouncil(!voteByCouncil)
+                  }}
+                />
+              )}
+
+              <Input
+                noMaxWidth
+                useDefaultStyle={false}
+                className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
+                wrapperClassName="mb-6"
+                label="Description"
+                placeholder="Describe your proposal (optional)"
+                value={form.description}
+                type="text"
+                onChange={(event) =>
+                  handleSetForm({
+                    value: event.target.value,
+                    propertyName: 'description',
+                  })
+                }
+              />
+            </>
+          )}
+
+          <div
+            className={`${
+              close ? 'justify-start' : 'justify-center'
+            } flex gap-x-6 items-center mt-8`}
+          >
+            <SecondaryButton
+              disabled={isLoading}
+              className="w-44"
+              onClick={() => close()}
+            >
+              Cancel
+            </SecondaryButton>
+
+            <Button
+              disabled={!form.destinationAccount}
+              className="w-44 flex justify-center items-center"
+              onClick={confirmPropose}
+              isLoading={isLoading}
+            >
+              Add member
+            </Button>
+          </div>
+        </>
+      ) : (
+        <div className="w-full flex md:flex-row flex-col justify-between items-start">
+          <div className="w-full flex md:mb-0 mb-20 flex-col gap-y-5 justify-start items-start md:max-w-xl rounded-xl">
+            <Input
+              useDefaultStyle={false}
+              className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
+              wrapperClassName="mb-6 w-full"
+              label="Member's wallet"
+              placeholder="Member's wallet"
+              value={form.destinationAccount}
+              type="text"
+              onChange={(event) =>
+                handleSetForm({
+                  value: event.target.value,
+                  propertyName: 'destinationAccount',
+                })
+              }
+              noMaxWidth
+              error={formErrors['destinationAccount']}
+            />
+
+            <Input
+              noMaxWidth
+              useDefaultStyle={false}
+              className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
+              wrapperClassName="mb-6 w-full md:w-1/2"
+              min={mintMinAmount}
+              label="Voter weight"
+              value={form.amount}
+              type="number"
+              onChange={setAmount}
+              step={mintMinAmount}
+              error={formErrors['amount']}
+              onBlur={validateAmountOnBlur}
+            />
+
             <VoteBySwitch
+              disabled={!canChooseWhoVote}
               checked={voteByCouncil}
               onChange={() => {
                 setVoteByCouncil(!voteByCouncil)
               }}
             />
-          )}
 
-          <Input
-            noMaxWidth
-            useDefaultStyle={false}
-            className="p-4 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none"
-            wrapperClassName="mb-6"
-            label="Description"
-            placeholder="Describe your proposal (optional)"
-            value={form.description}
-            type="text"
-            onChange={(event) =>
-              handleSetForm({
-                value: event.target.value,
-                propertyName: 'description',
-              })
-            }
-          />
-        </>
+            <Button
+              disabled={!form.destinationAccount}
+              className="w-44 flex justify-center items-center mt-8"
+              onClick={confirmPropose}
+              isLoading={isLoading || formErrors['destinationAccount']}
+            >
+              Add member
+            </Button>
+          </div>
+
+          <div className="md:max-w-xs w-full">
+            <Input
+              noMaxWidth
+              useDefaultStyle
+              wrapperClassName="mb-6"
+              label="Title of your proposal"
+              placeholder="Title of your proposal (optional)"
+              value={form.title || proposalTitle}
+              type="text"
+              onChange={(event) =>
+                handleSetForm({
+                  value: event.target.value,
+                  propertyName: 'title',
+                })
+              }
+            />
+
+            <Input
+              noMaxWidth
+              useDefaultStyle
+              wrapperClassName="mb-20"
+              label="Description"
+              placeholder="Describe your proposal (optional)"
+              value={form.description}
+              type="text"
+              onChange={(event) =>
+                handleSetForm({
+                  value: event.target.value,
+                  propertyName: 'description',
+                })
+              }
+            />
+
+            <TokenBalanceCard />
+          </div>
+        </div>
       )}
-
-      <div
-        className={`${
-          close ? 'justify-start' : 'justify-center'
-        } flex gap-x-6 items-center mt-8`}
-      >
-        {close && (
-          <SecondaryButton
-            disabled={isLoading}
-            className="w-44"
-            onClick={() => close()}
-          >
-            Cancel
-          </SecondaryButton>
-        )}
-
-        <Button
-          disabled={!form.destinationAccount}
-          className="w-44 flex justify-center items-center"
-          onClick={() => handlePropose()}
-          isLoading={isLoading}
-        >
-          Add member
-        </Button>
-      </div>
     </>
   )
 }
